+++ package.json
@@ -17,9 +17,11 @@
     "@headlessui/react": "^1.7.14",
     "next": "13.4.4",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "tailwindcss": "^3.3.2"
   },
   "devDependencies": {
+    "@types/node": "20.2.4",
     "@types/react": "18.2.7",
     "@typescript-eslint/eslint-plugin": "^5.59.7",
     "@typescript-eslint/parser": "^5.59.7",
